Moved more files to the globalComponents folder and created two subfolders
