+++ src/armory/Armory.js
@@ -9,22 +9,22 @@ import './Armory.css';
 import Navbar from '../globalComponents/Navbar.js';
 import CreateNewTankPopup from './CreateNewTankPopup.js';
 import DeleteTankPopup from './DeleteTankPopup.js';
-import SelectTank from './SelectTank.js';
+import SelectTank from '../globalComponents/SelectTank.js';
 import SetWagerPopup from './SetWagerPopup.js';
 import RenameTankPopup from './RenameTankPopup.js';
 // Functions
-import { getInventory, getComponentPoints } from './GetInventoryInfo.js';
-import { getUser } from '../globalComponents/userAPIIntegration.js';
-import { getAllUsersTanks, getFavoriteTank } from '../globalComponents/tankAPIIntegration.js';
+import { getInventory, getComponentPoints } from '../globalComponents/GetInventoryInfo.js';
+import { getUser } from '../globalComponents/apiCalls/userAPIIntegration.js';
+import { getAllUsersTanks, getFavoriteTank } from '../globalComponents/apiCalls/tankAPIIntegration.js';
 import { getTank, getEmptyCasusCode } from '../tanks/TankLoader.js';
 import { toTitleCase } from '../globalComponents/Utility.js';
 // Types and Classes
-import type { TankComponent } from './TankComponent.js';
+import type { TankComponent } from '../globalComponents/typesAndClasses/TankComponent.js';
 import { verifyLink } from '../globalComponents/verifyLink.js';
-import { verifyLogin } from '../globalComponents/verifyLogin.js';
+import { verifyLogin } from '../globalComponents/apiCalls/verifyLogin.js';
 import TankPart from '../tanks/TankPart.js';
 import BackendTank from '../tanks/BackendTank.js';
-import Component from './Component.js';
+import Component from '../globalComponents/typesAndClasses/Component.js';
 import Tank from '../tanks/Tank.js';
 import Chassis from '../tanks/Chassis.js';
 import Gun from '../tanks/Gun.js';
