+++ src/battlearena/BattleArena.js
@@ -8,16 +8,16 @@ import Leaderboard from '../globalComponents/Leaderboard.js';
 import SearchPlayers from './SearchPlayers.js';
 import ChallengePlayerPopup from './ChallengePlayerPopup.js';
 import { verifyLink } from '../globalComponents/verifyLink.js';
-import { verifyLogin } from '../globalComponents/verifyLogin.js';
+import { verifyLogin } from '../globalComponents/apiCalls/verifyLogin.js';
 import setReturnToFromBattlegroundLink from '../battleground/setReturnToFromBattlegroundLink.js';
-import SelectTank from '../armory/SelectTank.js';
+import SelectTank from '../globalComponents/SelectTank.js';
 import Tank from '../tanks/Tank.js';
-import { getAllUsersTanks } from '../globalComponents/tankAPIIntegration.js';
+import { getAllUsersTanks } from '../globalComponents/apiCalls/tankAPIIntegration.js';
 import TankDisplay from '../tanks/TankDisplay.js';
-import User from '../globalComponents/User.js';
-import prepareMatchAPICall from '../globalComponents/prepareMatchAPICall.js';
-import {setMatchForBattleground} from '../battleground/setTanksToFightInBattleground.js';
-import getReplayListAPICall from '../globalComponents/getReplayListAPICall.js';
+import User from '../globalComponents/typesAndClasses/User.js';
+import prepareMatchAPICall from '../globalComponents/apiCalls/prepareMatchAPICall.js';
+import { setMatchForBattleground } from '../battleground/setTanksToFightInBattleground.js';
+import getReplayListAPICall from '../globalComponents/apiCalls/getReplayListAPICall.js';
 import { ToastContainer , toast } from 'react-toastify';
 import Replays from './Replays.js';
