+++ src/globalComponents/Navbar.js
@@ -3,9 +3,9 @@
 import * as React from 'react';
 import { Link } from 'react-router-dom';
 import Cookies from 'universal-cookie';
-import type { LinkType } from './LinkType.js';
+import type { LinkType } from './typesAndClasses/LinkType.js';
 import { verifyLink } from './verifyLink.js';
-import getUserAPICall from './getUserAPICall.js';
+import getUserAPICall from './apiCalls/getUserAPICall.js';
 
 type Props = {
 	linkName: LinkType,
