+++ src/globalComponents/typesAndClasses/Component.js
