+++ src/globalComponents/typesAndClasses/LinkType.js
