+++ src/globalComponents/typesAndClasses/Replay.js
