+++ src/globalComponents/typesAndClasses/TankComponent.js
