+++ src/globalComponents/verifyLink.js
@@ -1,6 +1,6 @@
 //@flow strict
 
-import type { LinkType } from './LinkType.js';
+import type { LinkType } from './typesAndClasses/LinkType.js';
 
 // Short function that verifies linktype is being used correctly.
 function verifyLink(linkName: LinkType): LinkType {
