+++ src/marketplace/ListingsView.js
@@ -1,11 +1,10 @@
 //@flow strict
 import * as React from 'react';
-import {getComponentType, verifyComponent} from '../armory/GetInventoryInfo.js';
+import { getComponentType, verifyComponent } from '../globalComponents/GetInventoryInfo.js';
 import type { SellingType } from './SellingType.js';
-import {getUser} from '../globalComponents/userAPIIntegration.js';
+import {getUser} from '../globalComponents/apiCalls/userAPIIntegration.js';
 import SaleObject from './SaleObject.js';
 import { ToastContainer , toast } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.min.css';
 import { toTitleCase } from '../globalComponents/Utility.js';
 
 type Props = {|
